go.sh muestras mas cosas, check net no jode si de entrada no hay red (por ejemplo si estas por eth)

+++ scripts/check_net.sh
@@ -1,18 +1,24 @@
 #!/bin/bash -E
 # Check if comm with server is running, if not, kill cmd.
 
-# wait for cmd startup
-sleep 4
+# Check if we have working wireless connection
+# If not, then don't bother with this (may be using eth)
+x=`ping -c1 10.42.43.1 2>&1 | grep "100% packet loss"`
+if [ ! "$x" = "" ]; then
+    echo "No connection, detected, will not control cmd..."
+    exit
+fi
+sleep 5
 
 # check connections
 while(true);
 do
-    x=`ping -c3 10.42.43.1 2>&1 | grep "100% packet loss"`
+    x=`ping -c1 10.42.43.1 2>&1 | grep "100% packet loss"`
     if [ ! "$x" = "" ]; then
 	echo "check_net.sh: Connection lost!"
 	echo "Will kill cmd!"
 	kill -9 `pidof -s cmd`
 	exit
     fi
-    sleep 1
+    sleep 5
 done
